% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterWaters.R
\name{check.cluster.method}
\alias{check.cluster.method}
\title{Check Clustering Method}
\usage{
check.cluster.method(cluster.method)
}
\arguments{
\item{cluster.method}{The user defined clustering method for the
\code{\link[=ConservedWaters]{ConservedWaters()}} and \code{\link[=ConservedWaters.MDS]{ConservedWaters.MDS()}}.}
}
\value{
Correctly formatted clustering method or a stop error
}
\description{
Ensures the user provided clustering method is a valid choice.
}
\details{
A simple check and reformatting of the clustering method indicated
by the user in the \code{\link[=ConservedWaters]{ConservedWaters()}} and \code{\link[=ConservedWaters.MDS]{ConservedWaters.MDS()}}
parameters.
}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
}
