% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{names.resATs.nitro.neut}
\alias{names.resATs.nitro.neut}
\title{Neutral Nitrogen Residue-AtomType Names}
\format{An object of class \code{character} of length 24.}
\usage{
names.resATs.nitro.neut
}
\description{
Neutral nitrogen residue-atomtype names based on PDB atom naming
conventions.
}
\details{
These residue-atomtype names indicate nitrogen atoms with a neutral
charge.
}
\examples{
 names.resATs.nitro.neut
 # [1] "ALA N"   "ARG N"   "ARG NE"  "ASN N"   "ASN ND2" "ASP N"   "CYS N"
 # "GLN N"   "GLN NE2" "GLU N"   "GLY N"   "HIS N"   "ILE N"   "LEU N"   "LYS N"
 # [16] "MET N"   "PHE N"   "PRO N"   "SER N"   "THR N"   "TRP N"   "TRP NE1"
 # "TYR N"   "VAL N"

}
\seealso{
Other constants: \code{\link{names.backbone.atoms}},
  \code{\link{names.polar.atoms}},
  \code{\link{names.res.AtomTypes}},
  \code{\link{names.resATs.carb.sulf}},
  \code{\link{names.resATs.nitro.pos}},
  \code{\link{names.resATs.oxy.neg}},
  \code{\link{names.resATs.oxy.neut}},
  \code{\link{names.residues}},
  \code{\link{names.sidechain.atoms}},
  \code{\link{names.waters}}
}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
}
\keyword{datasets}
