% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{names.resATs.nitro.pos}
\alias{names.resATs.nitro.pos}
\title{Positive Nitrogen Residue-AtomType Names}
\format{An object of class \code{character} of length 5.}
\usage{
names.resATs.nitro.pos
}
\description{
Positive nitrogen residue-atomtype names based on PDB atom
naming conventions.
}
\details{
These residue-atomtype names indicate nitrogen atoms with a positive
charge.
}
\examples{
 names.resATs.nitro.pos
 # [1] "ARG NH1" "ARG NH2" "HIS ND1" "HIS NE2" "LYS NZ"

}
\seealso{
Other constants: \code{\link{names.backbone.atoms}},
  \code{\link{names.polar.atoms}},
  \code{\link{names.res.AtomTypes}},
  \code{\link{names.resATs.carb.sulf}},
  \code{\link{names.resATs.nitro.neut}},
  \code{\link{names.resATs.oxy.neg}},
  \code{\link{names.resATs.oxy.neut}},
  \code{\link{names.residues}},
  \code{\link{names.sidechain.atoms}},
  \code{\link{names.waters}}
}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
}
\keyword{datasets}
