% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{names.resATs.oxy.neg}
\alias{names.resATs.oxy.neg}
\title{Negative Oxygen Residue-AtomType Names}
\format{An object of class \code{character} of length 4.}
\usage{
names.resATs.oxy.neg
}
\description{
Negatvie oxygen residue-atomtype names based on PDB atom naming
conventions.
}
\details{
These residue-atomtype names indicate oxygen atoms with a negative
charge.
}
\examples{
 names.resATs.oxy.neg
 # [1] "ASP OD1" "ASP OD2" "GLU OE1" "GLU OE2"

}
\seealso{
Other constants: \code{\link{names.backbone.atoms}},
  \code{\link{names.polar.atoms}},
  \code{\link{names.res.AtomTypes}},
  \code{\link{names.resATs.carb.sulf}},
  \code{\link{names.resATs.nitro.neut}},
  \code{\link{names.resATs.nitro.pos}},
  \code{\link{names.resATs.oxy.neut}},
  \code{\link{names.residues}},
  \code{\link{names.sidechain.atoms}},
  \code{\link{names.waters}}
}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
}
\concept{constants}
\keyword{datasets}
