% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{names.resATs.oxy.neut}
\alias{names.resATs.oxy.neut}
\title{Neutral Oxygen Residue-AtomType Names}
\format{An object of class \code{character} of length 25.}
\usage{
names.resATs.oxy.neut
}
\description{
Neutral oxygen residue-atomtype names based on PDB atom naming
conventions.
}
\details{
These residue-atomtype names indicate oxygen atoms with a neutral
charge.
}
\examples{
 names.resATs.oxy.neut
 # [1] "ALA O"   "ARG O"   "ASN O"   "ASN OD1" "ASP O"   "CYS O"   "GLN O"
 # "GLN OE1" "GLU O"   "GLY O"   "HIS O"   "ILE O"   "LEU O"   "LYS O"   "MET O"
 # [16] "PHE O"   "PRO O"   "SER O"   "SER OG"  "THR O"   "THR OG1" "TRP O"
 # "TYR O"   "TYR OH"  "VAL O"

}
\seealso{
Other constants: \code{\link{names.backbone.atoms}},
  \code{\link{names.polar.atoms}},
  \code{\link{names.res.AtomTypes}},
  \code{\link{names.resATs.carb.sulf}},
  \code{\link{names.resATs.nitro.neut}},
  \code{\link{names.resATs.nitro.pos}},
  \code{\link{names.resATs.oxy.neg}},
  \code{\link{names.residues}},
  \code{\link{names.sidechain.atoms}},
  \code{\link{names.waters}}
}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
}
\keyword{datasets}
