% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{names.waters}
\alias{names.waters}
\title{Water Residue Names}
\format{An object of class \code{character} of length 3.}
\usage{
names.waters
}
\description{
Water residue names based on PDB naming conventions.
}
\details{
The three (3) letter abbreviation for the three (3) commonly used
abbreviations for water residues.
}
\examples{
  names.waters
  # [1] "HOH", "DOD", "WAT"

}
\seealso{
Other constants: \code{\link{names.backbone.atoms}},
  \code{\link{names.polar.atoms}},
  \code{\link{names.res.AtomTypes}},
  \code{\link{names.resATs.carb.sulf}},
  \code{\link{names.resATs.nitro.neut}},
  \code{\link{names.resATs.nitro.pos}},
  \code{\link{names.resATs.oxy.neg}},
  \code{\link{names.resATs.oxy.neut}},
  \code{\link{names.residues}},
  \code{\link{names.sidechain.atoms}}
}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
}
\keyword{datasets}
