% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{OccupancyBarplot.summ}
\alias{OccupancyBarplot.summ}
\title{Occupancy Summary Barplots}
\usage{
OccupancyBarplot.summ(data)
}
\arguments{
\item{data}{The results from the \code{\link[=CleanProteinStructures]{CleanProteinStructures()}} function. Will
use the binned occupancy data.}
}
\description{
Occupancy summary barplots for the PDB structures. The plots are
faceted and displays the binned occupancy values for all the structures.
The counts are presented on a \code{log10} scale.
}
\examples{
  \dontrun{
  OccupancyBarplot.summ(data)

  ##----- multiple pages
  library(ggforce)
  occ.barplots.summary <- OccupancyBarplot.summ(data)
  num.pages <- ceiling(nrow(data$occupancy.counts) / 10)

  pdf(file="multiple_pages.pdf", height=11, width=8.5)
  for (page in seq_len(num.pages)) {
    print(occ.barplots.summary +
          ggforce::facet_wrap_paginate(~PDBid,
                                       ncol = 2, nrow = 5, page = page) )
  }
  dev.off()
  }

}
\seealso{
Other plots: \code{\link{BoundWaterEnvPlots}},
  \code{\link{BoundWaterEnvSummaryPlot}},
  \code{\link{BvalueBarplot.summ}},
  \code{\link{BvalueBarplot}},
  \code{\link{ClusterSummaryPlots}},
  \code{\link{MobNormBvalEvalPlots}},
  \code{\link{MobilityBarplot.summ}},
  \code{\link{MobilityBarplot}},
  \code{\link{OccupancyBarplot}},
  \code{\link{nBvalueBarplot}},
  \code{\link{normBvalueBarplot.summ}}
}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
}
\concept{plots}
