% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_documentation.R
\name{PDB.1ecd}
\alias{PDB.1ecd}
\title{PDB Structure of Erythrocruorin}
\description{
Structure of erythrocruorin in different ligand states refined
at 1.4 A resolution.
}
\details{
The 3D structure of erythrocruorin in different ligand states
refined at 1.4 A resolution. This 3D structure was downloaded from the
\href{http://www.rcsb.org/pdb/explore/explore.do?structureId=1ecd}{RCSB} and
read into \code{R} using \code{\link[bio3d:read.pdb]{bio3d::read.pdb()}}. It is used in examples and
testing.
}
\references{
PDB ID: \href{http://www.rcsb.org/pdb/explore/explore.do?structureId=1ecd}{1ecd}
Wolfgang Steigemann and Ernst Weber. Structure of erythrocruorin in different
ligand states refined at 1.4 A resolution. \emph{Journal of Molecular Biology},
1979, \strong{127} (\emph{3}), pp 309-338.
\href{https://doi.org/10.1016/0022-2836(79)90332-2}{DOI: 10.1016/0022-2836(79)90332-2}
\href{https://www.ncbi.nlm.nih.gov/pubmed/?term=430568}{PMID: 430568}

HM Berman, J Westbrook, Z Feng, G Gilliland, TN Bhat, H Weissig, IN
Shindyalov, PE Bourne. The Protein Data Bank. \emph{Nucleic Acids Research},
2000, \strong{28} (\emph{1}), pp 235-242.
\href{https://doi.org/10.1093/nar/28.1.235}{DOI: 10.1093/nar/28.1.235}
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC102472/}{PMCID: PMC102472}
}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
}
