% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_documentation.R
\name{PDB.5rxn}
\alias{PDB.5rxn}
\title{PDB Structure of Rubredoxin}
\description{
Combined crystallographic refinement and energy minimization of
rubredoxin at 1.2 Angstrom resolution.
}
\details{
The 3D structure of rubredoxin at 1.2 Angstrom resolution obtained
via combined crystallographic refinement and energy minimization. This 3D
structure was downloaded from the
\href{http://www.rcsb.org/pdb/explore/explore.do?structureId=5rxn}{RCSB} and
read into \code{R} using \code{\link[bio3d:read.pdb]{bio3d::read.pdb()}}. It is used in examples and
testing.
}
\references{
PDB ID: \href{http://www.rcsb.org/pdb/explore/explore.do?structureId=5rxn}{5rxn}
Keith D Watenpaugh. Combined Crystallographic Refinement And Energy
Minimization Of Rubredoxin At 1.2 Angstrom Resolution.

HM Berman, J Westbrook, Z Feng, G Gilliland, TN Bhat, H Weissig, IN
Shindyalov, PE Bourne. The Protein Data Bank. \emph{Nucleic Acids Research},
2000, \strong{28} (\emph{1}), pp 235-242.
\href{https://doi.org/10.1093/nar/28.1.235}{DOI: 10.1093/nar/28.1.235}
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC102472/}{PMCID: PMC102472}
}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
}
