% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_documentation.R
\name{thrombin.1hai}
\alias{thrombin.1hai}
\title{PDB Structure of Thrombin}
\description{
Isomorphous structures of prethrombin2, hirugen-, and
PPACK-thrombin.
}
\details{
The 3D structure of isomorphous structures of prethrombin2,
hirugen-, and PPACK-thrombin at 2.4 Angstroms. This 3D structure was
downloaded from the
\href{http://www.rcsb.org/pdb/explore/explore.do?structureId=1hai}{RCSB} and
read into \code{R} using \code{\link[bio3d:read.pdb]{bio3d::read.pdb()}}. It is used in examples and
testing.
}
\references{
PDB ID: \href{http://www.rcsb.org/pdb/explore/explore.do?structureId=1hai}{1hai}
J Vijayalakshmi, KP Padmanabhan, KG Mann, and A Tulinsky. The isomorphous
structures of prethrombin2, hirugen-, and PPACK-thrombin: changes
accompanying activation and exosite binding to thrombin.. \emph{Protein
Science}, 1994, \strong{3} (\emph{12}), pp 2254–2271.

HM Berman, J Westbrook, Z Feng, G Gilliland, TN Bhat, H Weissig, IN
Shindyalov, PE Bourne. The Protein Data Bank. \emph{Nucleic Acids Research},
2000, \strong{28} (\emph{1}), pp 235-242.
\href{https://dx.doi.org/10.1002/pro.5560031211}{DOI: 10.1002/pro.5560031211}
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2142772}{PMCID: PMC2142772}
}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
}
