% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_documentation.R
\name{thrombin10.PDBs.align}
\alias{thrombin10.PDBs.align}
\title{Thrombin10 Vignette's Primary Sequence Alignment}
\description{
Thrombin10 vignette's primary sequence alignment.
}
\details{
The primary sequence alignment is being provided because the CRAN
servers where R packages are tested does not have the MUltiple Sequence
Comparison by Log-Expectation (MUSCLE; \href{http://www.drive5.com/muscle/}{MUSCLEwebpage} and the \href{http://www.ebi.ac.uk/Tools/msa/muscle/}{EBIwebpage} application installed.
This alignment of 10 thrombin structures for the Thrombin 10 Vignette
allows the vignette to be completed without errors.
}
\references{
RC Edgar. MUSCLE: multiple sequence alignment with high accuracy
and high throughput. \emph{Nucleic Acids Research}, 2004, \strong{32} (\emph{5}),
pp 1792-1797. \cr
\href{http://www.ncbi.nlm.nih.gov/pubmed/15034147}{PMID: 15034147} \cr
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC390337/}{PMCID: PMC390337} \cr
\href{https://doi.org/10.1093/nar/gkh340}{DOI: 10.1093/nar/gkh340}

RC Edgar. MUSCLE: a multiple sequence alignment method with reduced time
and space complexity \emph{BMC Bioinformatics}, 2004, 5:113. \cr
\href{https://www.ncbi.nlm.nih.gov/pubmed/15318951}{PMID: 15318951} \cr
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC517706/}{PMCID: PMC517706} \cr
\href{https://doi.org/10.1186/1471-2105-5-113}{DOI: 10.1186/1471-2105-5-113}
}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
}
