% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanddraabe.R
\docType{package}
\name{vanddraabe}
\alias{vanddraabe}
\alias{vanddraabe-package}
\title{vanddraabe: Identification and Statistical Analysis of Conserved Waters in Proteins}
\description{
Identify and analyze conserved waters within crystallographic protein
structures and molecular dynamics simulation trajectories. Statistical
parameters for each water cluster, informative graphs, and a PyMOL session
file to visually explore the conserved waters and protein are returned.
Hydrophilicity is the propensity of waters to congregate near specific
protein atoms and is related to conserved waters. An informatics derived set
of hydrophilicity values are provided based on a large, high-quality X-ray
protein structure dataset.
}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
}
